remove tests/util
